feat: BaseGuildEmojiManager (#4934)
