+++ esm/discord.mjs
@@ -28,6 +28,7 @@ export const {
   UserFlags,
   Util,
   version,
+  BaseGuildEmojiManager,
   ChannelManager,
   GuildChannelManager,
   GuildEmojiManager,
